+++ app/src/container/MapContainer/index.js
@@ -3,8 +3,8 @@ import * as React from 'react';
 import {Alert, Platform} from 'react-native';
 import {connect} from 'react-redux';
 import Map from '../../screens/Map';
-import {fetchParkspots, filterParkspots, updateLocation, updateMapPosition} from './actions';
-import {clearSelectedLocation} from '../SearchContainer/actions';
+import {addFavoriteByDescription, fetchParkspots, filterParkspots, updateLocation, updateMapPosition} from './actions';
+import {clearSelectedLocation, remFavorite} from '../SearchContainer/actions';
 
 export interface Props {
   navigation: any;
@@ -18,6 +18,9 @@ export interface Props {
   filterParkspots: Function;
   distanceFilterValue: Number;
   clearSelectedLocation: Function;
+  addFavoriteByDescription: Function;
+  remFavorite: Function;
+  favorites: Object;
 }
 
 export interface State {
@@ -38,6 +41,9 @@ class MapContainer extends React.Component<Props, State> {
         filterParkspots={this.props.filterParkspots}
         distanceFilterValue={this.props.distanceFilterValue}
         clearSelectedLocation={this.props.clearSelectedLocation}
+        addFavoriteByDescription={this.props.addFavoriteByDescription}
+        remFavorite={this.props.remFavorite}
+        favorites={this.props.favorites}
       />
     );
   }
@@ -58,6 +64,8 @@ function bindAction(dispatch) {
     updateMapPosition: (mapPosition) => dispatch(updateMapPosition(mapPosition)),
     filterParkspots: (filterId) => dispatch(filterParkspots(filterId)),
     clearSelectedLocation: () => dispatch(clearSelectedLocation()),
+    addFavoriteByDescription: (description) => dispatch(addFavoriteByDescription(description)),
+    remFavorite: (fav) => dispatch(remFavorite(fav)),
   };
 }
 
@@ -67,5 +75,6 @@ const mapStateToProps = state => ({
   mapPosition: state.mapReducer.mapPosition,
   selectedLocation: state.searchReducer.selectedLocation,
   distanceFilterValue: state.filterReducer.distance,
+  favorites: state.searchReducer.favorites,
 });
 export default connect(mapStateToProps, bindAction)(MapContainer);